steamshim: deduplicate pipe implementation
