+++ source/steamshim/pipe.c
@@ -1,0 +1,80 @@
+#include "pipe.h"
+
+#ifdef _WIN32
+
+static int pipeReady(PipeType fd)
+{
+    DWORD avail = 0;
+    return (PeekNamedPipe(fd, NULL, 0, NULL, &avail, NULL) && (avail > 0));
+} /* pipeReady */
+
+static int writePipe(PipeType fd, const void *buf, const unsigned int _len)
+{
+    const DWORD len = (DWORD) _len;
+    DWORD bw = 0;
+    return ((WriteFile(fd, buf, len, &bw, NULL) != 0) && (bw == len));
+} /* writePipe */
+
+static int readPipe(PipeType fd, void *buf, const unsigned int _len)
+{
+    const DWORD len = (DWORD) _len;
+    DWORD br = 0;
+    return ReadFile(fd, buf, len, &br, NULL) ? (int) br : -1;
+} /* readPipe */
+
+static void closePipe(PipeType fd)
+{
+    CloseHandle(fd);
+} /* closePipe */
+
+static char *getEnvVar(const char *key, char *buf, const size_t _buflen)
+{
+    const DWORD buflen = (DWORD) _buflen;
+    const DWORD rc = GetEnvironmentVariableA(key, buf, buflen);
+    /* rc doesn't count null char, hence "<". */
+    return ((rc > 0) && (rc < buflen)) ? buf : NULL;
+} /* getEnvVar */
+
+#else
+
+static int pipeReady(PipeType fd)
+{
+    int rc;
+    struct pollfd pfd = { fd, POLLIN | POLLERR | POLLHUP, 0 };
+    while (((rc = poll(&pfd, 1, 0)) == -1) && (errno == EINTR)) { /*spin*/ }
+    return (rc == 1);
+} /* pipeReady */
+
+static int writePipe(PipeType fd, const void *buf, const unsigned int _len)
+{
+    const ssize_t len = (ssize_t) _len;
+    ssize_t bw;
+    while (((bw = write(fd, buf, len)) == -1) && (errno == EINTR)) { /*spin*/ }
+    return (bw == len);
+} /* writePipe */
+
+static int readPipe(PipeType fd, void *buf, const unsigned int _len)
+{
+    const ssize_t len = (ssize_t) _len;
+    ssize_t br;
+    while (((br = read(fd, buf, len)) == -1) && (errno == EINTR)) { /*spin*/ }
+    return (int) br;
+} /* readPipe */
+
+static void closePipe(PipeType fd)
+{
+    close(fd);
+} /* closePipe */
+
+static char *getEnvVar(const char *key, char *buf, const size_t buflen)
+{
+    const char *envr = getenv(key);
+    if (!envr || (strlen(envr) >= buflen))
+        return NULL;
+    strcpy(buf, envr);
+    return buf;
+} /* getEnvVar */
+
+#endif
+
+
