+++ source/steamshim/pipe.h
@@ -1,0 +1,34 @@
+#ifndef _INCL_STEAMSHIM_PIPE_H_
+#define _INCL_STEAMSHIM_PIPE_H_
+
+#ifdef _WIN32
+#define WIN32_LEAN_AND_MEAN 1
+#include <windows.h>
+typedef HANDLE PipeType;
+#define NULLPIPE NULL
+typedef unsigned __int8 uint8;
+typedef __int32 int32;
+typedef unsigned __int64 uint64;
+#include <stdio.h>
+#else
+#include <stdio.h>
+#include <stdlib.h>
+#include <string.h>
+#include <stdint.h>
+#include <unistd.h>
+#include <errno.h>
+#include <poll.h>
+#include <signal.h>
+// typedef uint8_t uint8;
+// typedef int32_t int32;
+// typedef uint64_t uint64;
+typedef int PipeType;
+#define NULLPIPE -1
+#endif
+
+static int writePipe(PipeType fd, const void *buf, const unsigned int _len);
+static int readPipe(PipeType fd, void *buf, const unsigned int _len);
+static void closePipe(PipeType fd);
+static char *getEnvVar(const char *key, char *buf, const size_t buflen);
+static int pipeReady(PipeType fd);
+#endif
