+++ source/steamshim/steamshim_local.h
@@ -1,0 +1,34 @@
+#ifndef _INCL_STEAMSHIM_COMMON_H_
+#define _INCL_STEAMSHIM_COMMON_H_
+typedef enum ShimCmd
+{
+    SHIMCMD_BYE,
+    SHIMCMD_PUMP,
+    SHIMCMD_REQUESTSTATS,
+    SHIMCMD_STORESTATS,
+    SHIMCMD_SETACHIEVEMENT,
+    SHIMCMD_GETACHIEVEMENT,
+    SHIMCMD_RESETSTATS,
+    SHIMCMD_SETSTATI,
+    SHIMCMD_GETSTATI,
+    SHIMCMD_SETSTATF,
+    SHIMCMD_GETSTATF,
+    SHIMCMD_GETSTEAMID,
+    SHIMCMD_GETPERSONANAME,
+} ShimCmd;
+typedef enum ShimEvent
+{
+    SHIMEVENT_BYE,
+    SHIMEVENT_STATSRECEIVED,
+    SHIMEVENT_STATSSTORED,
+    SHIMEVENT_SETACHIEVEMENT,
+    SHIMEVENT_GETACHIEVEMENT,
+    SHIMEVENT_RESETSTATS,
+    SHIMEVENT_SETSTATI,
+    SHIMEVENT_GETSTATI,
+    SHIMEVENT_SETSTATF,
+    SHIMEVENT_GETSTATF,
+    SHIMEVENT_GETSTEAMID,
+    SHIMEVENT_GETPERSONANAME,
+} ShimEvent;
+#endif
