+++ source/steamshim/steamshim_parent.cpp
@@ -23,6 +23,7 @@ typedef int PipeType;
 #endif
 
 #include "steam/steam_api.h"
+#include "pipe.c"
 
 #define DEBUGPIPE 0
 #if DEBUGPIPE
@@ -36,11 +37,8 @@ static int mainline(void);
 
 /* Windows and Unix implementations of this stuff below. */
 static void fail(const char *err);
-static bool writePipe(PipeType fd, const void *buf, const unsigned int _len);
-static int readPipe(PipeType fd, void *buf, const unsigned int _len);
 static bool createPipes(PipeType *pPipeParentRead, PipeType *pPipeParentWrite,
                         PipeType *pPipeChildRead, PipeType *pPipeChildWrite);
-static void closePipe(PipeType fd);
 static bool setEnvVar(const char *key, const char *val);
 static bool launchChild(ProcessType *pid);
 static int closeProcess(ProcessType *pid);
@@ -51,20 +49,6 @@ static void fail(const char *err)
     MessageBoxA(NULL, err, "ERROR", MB_ICONERROR | MB_OK);
     ExitProcess(1);
 } // fail
-
-static bool writePipe(PipeType fd, const void *buf, const unsigned int _len)
-{
-    const DWORD len = (DWORD) _len;
-    DWORD bw = 0;
-    return ((WriteFile(fd, buf, len, &bw, NULL) != 0) && (bw == len));
-} // writePipe
-
-static int readPipe(PipeType fd, void *buf, const unsigned int _len)
-{
-    const DWORD len = (DWORD) _len;
-    DWORD br = 0;
-    return ReadFile(fd, buf, len, &br, NULL) ? (int) br : -1;
-} // readPipe
 
 static bool createPipes(PipeType *pPipeParentRead, PipeType *pPipeParentWrite,
                         PipeType *pPipeChildRead, PipeType *pPipeChildWrite)
@@ -89,11 +73,6 @@ static bool createPipes(PipeType *pPipeParentRead, PipeType *pPipeParentWrite,
 
     return 1;
 } // createPipes
-
-static void closePipe(PipeType fd)
-{
-    CloseHandle(fd);
-} // closePipe
 
 static bool setEnvVar(const char *key, const char *val)
 {
@@ -137,22 +116,6 @@ static void fail(const char *err)
     _exit(1);
 } // fail
 
-static bool writePipe(PipeType fd, const void *buf, const unsigned int _len)
-{
-    const ssize_t len = (ssize_t) _len;
-    ssize_t bw;
-    while (((bw = write(fd, buf, len)) == -1) && (errno == EINTR)) { /*spin*/ }
-    return (bw == len);
-} // writePipe
-
-static int readPipe(PipeType fd, void *buf, const unsigned int _len)
-{
-    const ssize_t len = (ssize_t) _len;
-    ssize_t br;
-    while (((br = read(fd, buf, len)) == -1) && (errno == EINTR)) { /*spin*/ }
-    return (int) br;
-} // readPipe
-
 static bool createPipes(PipeType *pPipeParentRead, PipeType *pPipeParentWrite,
                         PipeType *pPipeChildRead, PipeType *pPipeChildWrite)
 {
@@ -174,11 +137,6 @@ static bool createPipes(PipeType *pPipeParentRead, PipeType *pPipeParentWrite,
 
     return 1;
 } // createPipes
-
-static void closePipe(PipeType fd)
-{
-    close(fd);
-} // closePipe
 
 static bool setEnvVar(const char *key, const char *val)
 {
